requirements.txt
rumps==0.2.2
python-vlc==1.1.2
requests==2.20.0
